requirements.txt
Django==5.0.6
djangorestframework
channels
channels-redis
daphne
gunicorn
mysqlclient
python-dotenv
drf-yasg
django-redis